requirements.txt
flask
werkzeug
PyPDF2
gtts
edge-tts
deep-translator
groq
python-dotenv